+++ meta/3rd/OpenResty/library/table.clear.lua
@@ -1,7 +1,0 @@
----@meta
-
---- Clear a table of its contents.
----@param t table
-local function clear(t) end
-
-return clear